redux transformating signup
