+++ client/src/pages/SignupPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import AuthTemplate from '../components/auth/AuthTemplate';
-import Signup from '../components/auth/SignupForm';
+import Signup from '../containers/auth/Signup';
 
 const RegisterPage = () => {
   return (
